WIP: driftctl scan: split output format and destination
**Early WIP, do not use or review**

The `driftctl scan` flag `--output` used to combine format and
destination into pseudo-URLs, e.g.: `console://stdout`, `json://a-file`.

It isn't really necessary for `driftctl scan` to write to anywhere other
than stdout or stderr. Most of the time, that is the desired destination
of the output, and pipes and redirects can be used when that's not the
case.

Removing the concept of output has a couple of advantages: this is a net
deletion / simplification of code, and we lose the  non-standard
pseudo-URL syntax, including the pseudo-filename "stdout".

All currently supported formats can be accessed using the new `--format`
flag, which defaults to "console".

Contrast this with the `gen-driftignore` subcommand - a very common use
case of this command is to append drift to `.driftignore`, so that is
the default destination, rather than stdout. In this case the idiom
`--output -` is supported to write to stdout, similar to `tar`.